Fix leading non-digit characters being removed from raw version
The raw version needs to be what the user passed in and before this commit, the removal of leading non-digit characters was being applied to the raw version, despite that part only being necessary for the internal representation.

This fix therefore stores the version being worked on as a separate variable so that the raw version that gets stored only has whitespace trimmed.

A regression test case has also been added to check that the input raw version is the same as the parsed version's display implementation.

Related issue - https://github.com/russellbanks/Komac/issues/1274